requirements.txt
numpy==1.24.3
pandas==2.0.2
openpyxl==3.1.2
xlsxwriter==3.1.0
selenium==4.11.2
tqdm==4.65.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
tensorflow==2.10.1  # 可选，GPU加速
protobuf==3.20.0
webdriver-manager==4.0.0
argparse==1.4.0
matplotlib==3.7.1
scipy==1.10.1 